+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added `rules.withoutTrailingSemicolon` metric
+- Added `comments.sizeRatioPercent` metric
+- Added `selectors.sizeRatioPercent` metric
+- Added `declarations.sizeRatioPercent` metric
+- Added `dataUris.sizeRatioPercent` metric
 
 ### Changed